+++ src/napari_matplotlib/tests/test_histogram.py
@@ -11,6 +11,22 @@ from napari_matplotlib.tests.helpers import (
 
 
 @pytest.mark.mpl_image_compare
+def test_histogram_2D_bins(make_napari_viewer, astronaut_data):
+    viewer = make_napari_viewer()
+    viewer.theme = "light"
+    viewer.add_image(astronaut_data[0], **astronaut_data[1])
+    widget = HistogramWidget(viewer)
+    viewer.window.add_dock_widget(widget)
+    widget.bins_start = 0
+    widget.bins_stop = 350
+    widget.bins_num = 35
+    fig = widget.figure
+    # Need to return a copy, as original figure is too eagerley garbage
+    # collected by the widget
+    return deepcopy(fig)
+
+
+@pytest.mark.mpl_image_compare
 def test_histogram_2D(make_napari_viewer, astronaut_data):
     viewer = make_napari_viewer()
     viewer.theme = "light"